requirement.txt
PyYAML==6.0
openapi-core==0.17.1
Jinja2==3.1.2